R Source Code
I developed the code in 3 hours the implements a MACD momentum strategy and computes in my Market Microstructure and Trading Strategies course final exam. The program computes a MACD strategy based on the investor's risk appetite, computes the risk adjusted Sharpe ratio and a P&L statement for the trading strategy. 

Note: The detailed problem statement can be viewed in the next file